Fix incorrect rendering of transparent fill color for 2-sided pm3d surfaces
In the implementation of "with pm3d fc ls N", the color information
(stored as a 32-bit signed integer in the linestyle) is retrieved as follows:
It is first cast to a double, and then back to a 32-bit unsigned int
during processing.
For example, the color #8000ff00 is stored as -2147418368 (32-bit signed int).
Casting this to double gives -2147418368.0, and casting that to a 32-bit
unsigned int should result in 2147548928 (0x8000ff00), which represents a
semi-transparent green.
However, on ARM systems, casting a negative double -2147418368.0
to a 32-bit unsigned int yields 0, resulting in the default rendering
color of opaque black.

Bug 2813

+++ src/pm3d.c
@@ -732,7 +732,7 @@ pm3d_plot(struct surface_points *this_plot, int at_which_z)
 			struct lp_style_type style;
 			int side = pm3d_side( &pointsA[i], &pointsA[i1], &pointsB[ii]);
 			lp_use_properties(&style, side < 0 ? cb1 + 1 : cb1);
-			cb1 = cb2 = cb3 = cb4 = style.pm3d_color.lt;
+			cb1 = cb2 = cb3 = cb4 = (unsigned int) style.pm3d_color.lt;
 		    }
 		} else {
 		    cb1 = pointsA[i].z;
